Need to be careful about the final state of the tests...
It turns out that in very extreme edge cases (like with testify), we can see test output that looks like a test failed, but in fact the overall package level passed.  Need to honor the state of the package level.

tl;dr

testify is a test assertion package, and to test certain things, it runs tests inside other tests.  The inner tests produce output, but don't change the pass/fail status of the outer test.