Fix the test output
Roll up all output to the package level, and dump only when the package is finished, so the test output is adjacent to the package out.  Otherwise, you don't know which package the test was in.

Also retain all the package level output and print it out.  It's maybe a little noisy, but simpler and looks more like normal test output.